Expose extension-compatible CoreAardvark
