+++ Aardvark/ARKLogging.m
@@ -20,24 +20,8 @@
 
 #import "ARKLogging.h"
 
-#import "ARKLogDistributor.h"
+#import "ARKLogDistributor+UIAdditions.h"
 
-
-void ARKLog(NSString *format, ...)
-{
-    va_list argList;
-    va_start(argList, format);
-    [[ARKLogDistributor defaultDistributor] logWithFormat:format arguments:argList];
-    va_end(argList);
-}
-
-void ARKLogWithType(ARKLogType type, NSDictionary *userInfo, NSString *format, ...)
-{
-    va_list argList;
-    va_start(argList, format);
-    [[ARKLogDistributor defaultDistributor] logWithType:type userInfo:userInfo format:format arguments:argList];
-    va_end(argList);
-}
 
 void ARKLogScreenshot()
 {
