+++ Categories/UIApplication+ARKAdditions.m
@@ -18,13 +18,17 @@
 //  limitations under the License.
 //
 
+#if COCOAPODS
+#import "ARKLogDistributor.h"
+#else
+#import <CoreAardvark/ARKLogDistributor.h>
+#endif
 #import <objc/runtime.h>
 
 #import "UIApplication+ARKAdditions.h"
 
 #import "AardvarkDefines.h"
 #import "ARKBugReporter.h"
-#import "ARKLogDistributor.h"
 
 
 @interface UIApplication (ARKAdditions_Private)
