+++ CoreAardvark/ARKCoreLogging.h
@@ -1,9 +1,9 @@
 //
-//  ARKLogDistributor_Protected.h
-//  Aardvark
+//  ARKCoreLogging.h
+//  CoreAardvark
 //
-//  Created by Dan Federman on 3/30/15.
-//  Copyright 2015 Square, Inc.
+//  Created by Dan Federman on 7/25/16.
+//  Copyright 2016 Square, Inc.
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -18,8 +18,15 @@
 //  limitations under the License.
 //
 
-@interface ARKLogDistributor (Protected)
+#if COCOAPODS
+#import <Aardvark/ARKLogTypes.h>
+#else
+#import <CoreAardvark/ARKLogTypes.h>
+#endif
 
-- (void)waitUntilAllPendingLogsHaveBeenDistributed;
 
-@end
+/// Logs a log with default type to the default log distributor.
+OBJC_EXTERN void ARKLog(NSString * __nonnull format, ...) NS_FORMAT_FUNCTION(1,2);
+
+/// Logs a log with customized type and userInfo to the default log distributor.
+OBJC_EXTERN void ARKLogWithType(ARKLogType type, NSDictionary * __nullable userInfo, NSString * __nonnull format, ...) NS_FORMAT_FUNCTION(3,4);
