+++ CoreAardvark/ARKLogTypes.h
@@ -1,9 +1,9 @@
 //
-//  ARKLogObserver.h
-//  Aardvark
+//  ARKLogTypes.h
+//  CoreAardvark
 //
-//  Created by Dan Federman on 10/8/14.
-//  Copyright 2014 Square, Inc.
+//  Created by Evan Kimia on 10/23/15.
+//  Copyright 2015 Square, Inc.
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -21,22 +21,13 @@
 #import <Foundation/Foundation.h>
 
 
-@class ARKLogDistributor;
-@class ARKLogMessage;
-
-
-NS_ASSUME_NONNULL_BEGIN
-
-
-@protocol ARKLogObserver <NSObject>
-
-/// The log distributor that distributes logs to this observer.
-@property (weak, nullable) ARKLogDistributor *logDistributor;
-
-/// Called on a background operation queue when logs are appended to the log distributor.
-- (void)observeLogMessage:(ARKLogMessage *)logMessage;
-
-@end
-
-
-NS_ASSUME_NONNULL_END
+typedef NS_ENUM(NSUInteger, ARKLogType) {
+    /// Default log type.
+    ARKLogTypeDefault,
+    /// Marks the beginning or end of a task.
+    ARKLogTypeSeparator,
+    /// Marks that the log represents an error.
+    ARKLogTypeError,
+    /// Marks a log that has a screenshot attached.
+    ARKLogTypeScreenshot,
+};
