+++ CoreCategories/NSFileHandle+ARKAdditions.h
@@ -1,6 +1,6 @@
 //
 //  NSFileHandle+ARKAdditions.h
-//  Aardvark
+//  CoreAardvark
 //
 //  Created by Peter Westen on 3/17/15.
 //  Copyright 2015 Square, Inc.
