+++ CoreLogging/ARKLogMessage.m
@@ -1,6 +1,6 @@
 //
 //  ARKLogMessage.m
-//  Aardvark
+//  CoreAardvark
 //
 //  Created by Dan Federman on 10/4/14.
 //  Copyright 2014 Square, Inc.
