+++ CoreLogging/ARKLogObserver.h
@@ -1,8 +1,8 @@
 //
-//  ARKLogStore_Testing.h
-//  Aardvark
+//  ARKLogObserver.h
+//  CoreAardvark
 //
-//  Created by Dan Federman on 11/13/14.
+//  Created by Dan Federman on 10/8/14.
 //  Copyright 2014 Square, Inc.
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
@@ -18,11 +18,25 @@
 //  limitations under the License.
 //
 
-@class ARKDataArchive;
+#import <Foundation/Foundation.h>
 
 
-@interface ARKLogStore (Private)
+@class ARKLogDistributor;
+@class ARKLogMessage;
 
-@property ARKDataArchive *dataArchive;
+
+NS_ASSUME_NONNULL_BEGIN
+
+
+@protocol ARKLogObserver <NSObject>
+
+/// The log distributor that distributes logs to this observer.
+@property (weak, nullable) ARKLogDistributor *logDistributor;
+
+/// Called on a background operation queue when logs are appended to the log distributor.
+- (void)observeLogMessage:(ARKLogMessage *)logMessage;
 
 @end
+
+
+NS_ASSUME_NONNULL_END
