+++ CoreOther/AardvarkDefines.h
