+++ Gemfile
@@ -1,3 +1,4 @@
-source 'https://rubygems.org'
-
-gem 'cocoapods', '0.39.0'
+source 'https://rubygems.org' do
+  gem 'cocoapods', '1.0.0'
+  gem 'activesupport', '~> 4.0'
+end
