+++ Log Viewing/ARKIndividualLogViewController.m
@@ -16,12 +16,15 @@
 //  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
-
 //
 
-#import "ARKIndividualLogViewController.h"
-
+#if COCOAPODS
 #import "ARKLogMessage.h"
+#else
+#import <CoreAardvark/ARKLogMessage.h>
+#endif
+
+#import "ARKIndividualLogViewController.h"
 
 
 @interface ARKIndividualLogViewController ()
